requirements.txt
psycopg2-binary
python-dotenv
PyGithub
openai
flake8
bandit
slack-sdk
